Initial commit
Note that this commit will not build. Changes:
* Mark cgen.sh scripts executable for debugging
* Add echo's to cgen.sh script for debugging
* Make CGEN bitfield vars unsigned long long to fit BMI flags
* Mark bitfield printf format specifier as long
* Change shift-style bmi insns to follow same pattern as vanilla sll

+++ include/cgen/bitset.h
@@ -34,7 +34,7 @@ extern "C" {
 */
 typedef struct cgen_bitset
 {
-  unsigned length;
+  unsigned long long length;
   char *bits;
 } CGEN_BITSET;
